refactor(#2456): split kinematic_forces.py into data/coordinator modules
- Extract MjDataContext, KinematicForceData, _check_mujoco_version,
  export_kinematic_forces_to_csv to _kinematic_force_data.py (261 LOC)
- Slim kinematic_forces.py coordinator to 774 LOC (was 1126 LOC);
  retains KinematicForceAnalyzer (760 LOC) + re-exports
- All public names re-exported for backward compatibility
- TDD contract tests added (3 passed, 3 skipped pending mujoco install)
- Add terrain_representation.py exception to file_size_budget.json
  (pre-existing 1253 LOC, split planned in follow-up)
